[ADD] This is a change request to a previous ticket.
The original ticket created the analytic_account fields in the crm_team, payment_method and res_users models.
This change request sets the analytic_distribution on the account.move.lines when they are created or updated based on these fields.

[IMP] Improved efficiency of creation of analytic.distribution

[IMP] Removed write function.

The client should not modify the analytic_distribution but we do not want to limit them in case a user changes from one zone to another but forgets to update the res.user fields.
The setting of analytic_distribution now only need to take place on the create.