[ADD][3776834] jasman_sale_template: Multiple templates per Sale Order (#25)
The client typically sells product packages, consisting mostly of services rather than storable products. As a result, they prefer not to use Manufacturing kits. To accommodate this, they requested the ability to add multiple quotation templates to a single order, enabling them to utilize them as packages. When a template is linked to a sale order, all of its lines should be added to the order, plus a section with the name of the template.